Split "multiple records" into separate action
