+++ plugins/action_modify_multiple_records/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "modify_multiple_records_action",
+  "version": "0.1",
+  "description": "Inserts or updates multiple database records",
+  "author": "The mSupplyFoundation",
+  "license": "MIT",
+  "scripts": {
+    "build": "node ../../utils/pluginScripts/buildPluginFromPluginFolder.js",
+    "test": "node ../../utils/pluginScripts/testPluginFromPluginFolder.js"
+  },
+  "devDependencies": {},
+  "dependencies": {}
+}
