+++ plugins/action_modify_multiple_records/src/index.ts
@@ -1,0 +1,4 @@
+import { ActionPluginType } from '../../types'
+import modifyMultipleRecords from './modifyMultipleRecords'
+const action: ActionPluginType = modifyMultipleRecords
+export { action }
